Move TimeScheduler to dedicated module
