+++ gamdpy/runtime_actions/__init__.py
@@ -1,7 +1,8 @@
 # Runtime actions
 from .runtime_action import RuntimeAction, add_runtime_actions_list
 from .scalar_saver import ScalarSaver
-from .trajectory_saver import TrajectorySaver, TimeScheduler
+from .trajectory_saver import TrajectorySaver
+from .time_scheduler import TimeScheduler
 from .restart_saver import RestartSaver
 from .momentum_reset import MomentumReset
 from .stress_saver import StressSaver, extract_stress_tensor
